### 2018-04-06 #664 @whedon
[PRE REVIEW]: ACTIN: A tool to calculate stellar activity indices
**Submitting author:** @gomesdasilva (<a href="http://orcid.org/0000-0001-8056-9202">João Gomes da Silva</a>)
**Repository:** <a href="https://github.com/gomesdasilva/ACTIN" target ="_blank">https://github.com/gomesdasilva/ACTIN</a>
**Version:** v1.0
