### 2018-04-06 #664 @arfon
@gomesdasilva - please fix up the YAML header in your `paper.md` (I think the date field is currently commented out).

Also, do you have any suggested reviewers for this submission?
