### 2018-04-06 #664 @arfon
:wave: @RuthAngus @lmwalkowicz - would you be willing to review this submission for JOSS?
