### 2018-04-30 #667 @RuthAngus
@arfon I submitted some issues for ACTIN almost a week ago but I am not sure whether the author has seen these. Should I try to get in touch with the author another way?
