### 2018-06-15 #667 @arfon
How are we getting along here @gomesdasilva and @RuthAngus?
