+++ paper.bib
@@ -5,14 +5,9 @@ year = {2011},
 doi = {10.1051/0004-6361/201116971}
 }
 
-howpublished = {\url{https://doi.org/10.1051/0004-6361/201116971}}
-
-
 @unpublished{delgadomena2018,
 author = {{Delgado Mena}, E. and {Lovis} C., and {Santos} N.~C., and {Gomes da Silva} J., and {Mortier} A., and {Tsantaki} M., and {Sousa} S.~G., and {Figueira} P., and {Cunha} M.~S., and {Campante} T.~L., and {Adibekyan} V., and {Faria} J.~P., and {Montalto} M.},
 title = "{Planets Around Evolved Intermediate-Mass Stars in Open Clusters II. Are There Really Planets Around IC4651No9122, NGC2423No3 and NGC4349No127?}",
 year = {2018},
 note = {submitted}
 }
-
-author = {{Gomes da Silva}, J. and {Santos}, N.~C. and {Bonfils}, X. and {Delfosse}, X. and {Forveille}, T. and {Udry}, S.},
