### 2018-07-11 dcf2a29
update for compatibility with python3
