+++ actin/actin_files/actin_read_e2ds.py
@@ -1,5 +1,9 @@
 #!/usr/bin/env python
 
+# compatibility with python 2/3:
+from __future__ import print_function
+from __future__ import division
+
 import os
 import glob
 import numpy as np
@@ -49,15 +53,15 @@ def check_for_calib_files(e2ds_header, file_type, folder, dif_time_max=1.0):
 			index = time_dif.index(min(time_dif))
 			calib_pfile = calib_pfiles[index]
 			calib = pyfits.open(calib_pfile) # test new file
-			print "New %s file used:\t%s" % (file_type, os.path.split(calib_pfile)[-1])
-			print "%s file time difference to e2ds = %.2f days" % (file_type,min(time_dif))
+			print("New %s file used:\t%s" % (file_type, os.path.split(calib_pfile)[-1]))
+			print("%s file time difference to e2ds = %.2f days" % (file_type,min(time_dif)))
 			calib.close()
 			return calib_pfile
 		else:
-			print "*** WARNING: Closest %s file was produced longer than 1 day" % file_type
+			print("*** WARNING: Closest %s file was produced longer than 1 day" % file_type)
 			return None
 	else:
-		print "*** WARNING: No more %s files in folder" % file_type
+		print("*** WARNING: No more %s files in folder" % file_type)
 		return None
 
 
@@ -145,16 +149,16 @@ def read_file_e2ds(e2ds_pfile, obj_name=None):
 		==========  ========================================================
 	"""
 
-	print "\nREADING FILES FROM e2ds FITS FILE"
-	print "--------------------------------"
+	print("\nREADING FILES FROM e2ds FITS FILE")
+	print("--------------------------------")
 
 	if not e2ds_pfile:
-		print "*** ERROR: e2ds file with path required"
+		print("*** ERROR: e2ds file with path required")
 		return
 
 	try: e2ds = pyfits.open(e2ds_pfile)
 	except:
-		print "*** ERROR: Cannot read %s" % e2ds_pfile
+		print("*** ERROR: Cannot read %s" % e2ds_pfile)
 		return
 
 	folder, e2ds_file = os.path.split(e2ds_pfile)
@@ -163,32 +167,32 @@ def read_file_e2ds(e2ds_pfile, obj_name=None):
 
 	e2ds_file_info = e2ds_file.split('_')[0]
 
-	print "WORKING FOLDER:\t%s%s" % (folder, os.path.sep)
-	print "READING FILE:\t%s" % e2ds_file
+	print("WORKING FOLDER:\t%s%s" % (folder, os.path.sep))
+	print("READING FILE:\t%s" % e2ds_file)
 
 	tel = e2ds[0].header['TELESCOP'] # ESO-3P6 for HARPS, TNG for HARPS-N
 	instr = e2ds[0].header['INSTRUME'] # instrument used
 
-	print "TELESCOPE:\t%s" % tel
-	print "INSTRUMENT:\t%s" % instr
+	print("TELESCOPE:\t%s" % tel)
+	print("INSTRUMENT:\t%s" % instr)
 
 	if instr == 'HARPS': obs = 'ESO'
 	elif instr == 'HARPN': obs = 'TNG'
 	else:
-		print "*** ERROR: Instrument not recognized"
+		print("*** ERROR: Instrument not recognized")
 
 	try: obj = e2ds[0].header['OBJECT']
 	except:
 		try: obj = e2ds[0].header['%s OBS TARG NAME' % obs]
 		except:
-			print "*** ERROR: Cannot identify object"
+			print("*** ERROR: Cannot identify object")
 			return
 
-	print "OBJECT:\t\t%s" % obj
+	print("OBJECT:\t\t%s" % obj)
 
[…]
