+++ setup.py
@@ -6,7 +6,7 @@ try: from setuptools import setup
 except: from distutils.core import setup
 
 setup(name = 'actin',
-      version = '1.0',
+      version = '1.1',
       description = 'Activity Indices Calculator',
       url = 'http://github.com/gomesdasilva/actin',
       download_url = 'https://github.com/gomesdasilva/ACTIN/archive/v1.0.tar.gz',
