### 2018-09-19 425e325
included matplotlib as required module in README
