### 2018-09-19 a866989
added example to run code with test files in README
