### 2018-09-20 28799bb
added os.path.join to contruct directories in actin.py
