+++ actin/actin.py
@@ -17,7 +17,9 @@ import appdirs
 path = os.path.dirname(os.path.realpath(__file__)) # directory of actin.py
 
 # location of ACTIN files:
-sys.path.append("%s/actin_files/" % path)
+actin_files_dir = os.path.join(path, "actin_files") ##
+sys.path.append(actin_files_dir) ##
+#sys.path.append("%s/actin_files/" % path)
 import actin_config
 import actin_read_e2ds as read_e2ds
 import actin_read_s1d as read_s1d
@@ -30,7 +32,8 @@ import actin_functions as func
 
 
 # Configuration file:
-config_file = '%s/config_lines.txt' % path
+#config_file = '%s/config_lines.txt' % path
+config_file = os.path.join(path, "config_lines.txt") ##
 
 
 def actin_file(file, calc_index, config_file=config_file, save_output=False, line_plots=False, obj_name=None, targ_list=None, del_out=False, weight=None, norm='npixels'):
