### 2018-09-20 cd20bd6
Added automatic tests
