### 2018-09-25 aabbadc
Update to frac_neg to calculate fraction of pixels with negative flux. Addad description of test option to README
