+++ README.md
@@ -59,6 +59,8 @@ Arguments:
 
 `-cfg` bool : If `True` gives the path to the configuration file. Default is `False`.
 
+`-t` str : Tests the program using the test fits files provided. Use `e2ds`, `s1d` or `adp` to test e2ds, s1d or ADP files, respectively.
+
 ### Testing the code with minimum arguments:
 
 The example below will test the code using the test files provided in the package.
