+++ actin.egg-info/SOURCES.txt
@@ -1,0 +1,53 @@
+MANIFEST.in
+README.md
+setup.cfg
+setup.py
+actin/.DS_Store
+actin/__init__.py
+actin/__main__.py
+actin/actin.py
+actin/config_lines.txt
+actin.egg-info/PKG-INFO
+actin.egg-info/SOURCES.txt
+actin.egg-info/dependency_links.txt
+actin.egg-info/entry_points.txt
+actin.egg-info/requires.txt
+actin.egg-info/top_level.txt
+actin/actin_files/.DS_Store
+actin/actin_files/actin_calc_ind.py
+actin/actin_files/actin_calc_ind.pyc
+actin/actin_files/actin_config.py
+actin/actin_files/actin_config.pyc
+actin/actin_files/actin_functions.py
+actin/actin_files/actin_functions.pyc
+actin/actin_files/actin_get_win.py
+actin/actin_files/actin_get_win.pyc
+actin/actin_files/actin_plot_time.py
+actin/actin_files/actin_plot_time.pyc
+actin/actin_files/actin_read_adp.py
+actin/actin_files/actin_read_adp.pyc
+actin/actin_files/actin_read_e2ds.py
+actin/actin_files/actin_read_e2ds.pyc
+actin/actin_files/actin_read_s1d.py
+actin/actin_files/actin_read_s1d.pyc
+actin/actin_files/actin_save_data.py
+actin/actin_files/actin_save_data.pyc
+actin/actin_files/__pycache__/actin_calc_ind.cpython-36.pyc
+actin/actin_files/__pycache__/actin_config.cpython-36.pyc
+actin/actin_files/__pycache__/actin_functions.cpython-36.pyc
+actin/actin_files/__pycache__/actin_get_win.cpython-36.pyc
+actin/actin_files/__pycache__/actin_plot_time.cpython-36.pyc
+actin/actin_files/__pycache__/actin_read_adp.cpython-36.pyc
+actin/actin_files/__pycache__/actin_read_e2ds.cpython-36.pyc
+actin/actin_files/__pycache__/actin_read_s1d.cpython-36.pyc
+actin/actin_files/__pycache__/actin_save_data.cpython-36.pyc
+actin/test_files/.DS_Store
+actin/test_files/ADP.2014-09-16T11:04:45.123.fits
+actin/test_files/HARPS.2003-12-12T21:42:45.093_blaze_A.fits
+actin/test_files/HARPS.2003-12-12T21:47:59.366_wave_A.fits
+actin/test_files/HARPS.2003-12-13T06:19:48.371_bis_M2_A.fits
+actin/test_files/HARPS.2003-12-13T06:19:48.371_ccf_M2_A.fits
+actin/test_files/HARPS.2003-12-13T06:19:48.371_e2ds_A.fits
+actin/test_files/HARPS.2003-12-13T06:19:48.371_s1d_A.fits
+actin/test_files/HARPS.2006-07-05T03:52:40.071_bis_G2_A.fits
+actin/test_files/HARPS.2006-07-05T03:52:40.071_ccf_G2_A.fits
