### 2018-10-09 f88300d
added contact address for issues and problems in readme file
