+++ README.md
@@ -75,4 +75,4 @@ NOTE: If not installed via pip, use `python actin.py` instead of `actin`.
 
 This will execute ACTIN for all the subdirectories inside `../fits/` with files ending with `e2ds_A.fits`, calculate the indices `I_CaII` and `I_Ha`, output the data to `../output/star_names`, save spectra of the line regions to the same directory as data, and, before running the code, delete any output file that was previously there, in this case `Gl273_e2ds_actin.rdb` and `Gl581_e2ds_actin.rdb` files. Only fits files belonging to the stars chosen in `-tl` will be read, in this case `Gl273` and `Gl581`. In this case, the flux is not weighted (-w None, default) and is normalised by the number of pixels in the passband (-n npixels, default).
 
-For issues, problems and support contact jntgds(at)gmail.com
+For issues, problems, support or if you would like to contribute to the development of this code please contact jntgds(at)gmail.com
