### 2018-10-26 #667 @arfon
> Letting you know, @arfon is currently OOO until Monday, October 29th 2018. ❤️

It's a lie. I'm here :-)

> Hi @arfon and @gomesdasilva I think this code has met all the criteria and is ready to be accepted! How should I proceed?

Great, thanks @RuthAngus!
