### 2018-11-02 #667 @arfon
> @gomesdasilva - At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.

Friendly reminder to make an archive of the software sometime soon @gomesdasilva. You can read how to do that here: https://guides.github.com/activities/citable-code/
