### 2018-11-06 0dc940e
Update README.md
Added DOI
