### 2018-11-06 #667 @gomesdasilva
Hi Arfon,

I uploaded the software to Zenodo, DOI: 10.5281/zenodo.1478610.

Regards,
João

> On 2 Nov 2018, at 01:11, Arfon Smith <notifications@github.com> wrote:
>
> @gomesdasilva <https://github.com/gomesdasilva> - At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
>
> Friendly reminder to make an archive of the software sometime soon @gomesdasilva <https://github.com/gomesdasilva>. You can read how to do that here: https://guides.github.com/activities/citable-code/ <https://guides.github.com/activities/citable-code/>
