### 2018-11-06 fd5540e
Footnote update
