### 2018-11-06 #667 @arfon
@RuthAngus - many thanks for your review here ✨

@gomesdasilva - your paper is now accepted into JOSS :zap: :rocket: :boom:
