### 2018-11-06 #667 @gomesdasilva
Thank you Arfon and Ruth!

Just one last thing, is it possible to change the name from “Da Silva et al.” to "Gomes da Silva et al.”?
The last one is the one I use in my publications.

Cheers,
João

> On 6 Nov 2018, at 13:32, Arfon Smith <notifications@github.com> wrote:
>
> @RuthAngus <https://github.com/RuthAngus> - many thanks for your review here ✨
>
> @gomesdasilva <https://github.com/gomesdasilva> - your paper is now accepted into JOSS ⚡️ 🚀 💥
>
