### 2018-11-06 #667 @arfon
> Just one last thing, is it possible to change the name from “Da Silva et al.” to "Gomes da Silva et al.”?
The last one is the one I use in my publications.

Sure thing. I've updated this now. The online PDF might take a few hours to update.
