### 2018-11-06 #667 @gomesdasilva
Thank you very much

Cheers,
João

> On 6 Nov 2018, at 14:04, Arfon Smith <notifications@github.com> wrote:
>
> Just one last thing, is it possible to change the name from “Da Silva et al.” to "Gomes da Silva et al.”?
> The last one is the one I use in my publications.
>
> Sure thing. I've updated this now. The online PDF might take a few hours to update.
>
